cfg->data len uses uint32_t

+++ firmware/common/phal_G4/spi/spi.h
@@ -41,7 +41,7 @@ typedef enum {
  */
 typedef struct {
     uint32_t data_rate;          /*!< Desired SPI clock frequency (Hz) */
-    uint8_t data_len;            /*!< Transfer width in bits */
+    uint32_t data_len;           /*!< Transfer width in bits */
     SPI_Mode mode;               /*!< Master or slave mode */
 
     bool nss_sw;                 /*!< true = software-controlled chip select */
